Move assertCompiles into assertions helper class
