+++ src/test/java/sorald/processor/ProcessorTestFilesCompileTest.java
@@ -1,20 +1,11 @@
 package sorald.processor;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.notNullValue;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static sorald.Assertions.assertCompiles;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.List;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import javax.tools.Diagnostic;
-import javax.tools.DiagnosticCollector;
-import javax.tools.JavaFileObject;
-import javax.tools.ToolProvider;
 import org.apache.commons.io.FileUtils;
-import org.apache.commons.lang3.tuple.Pair;
 import org.junit.jupiter.api.io.TempDir;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -25,8 +16,7 @@ class ProcessorTestFilesCompileTest {
 
     @ParameterizedTest
     @MethodSource("provideCompilableProcessorTestInputFile")
-    void processorTestCaseInputFile_notMarkedNOCOMPILE_shouldCompile(File testCaseJavaFile)
-            throws IOException {
+    void processorTestCaseInputFile_notMarkedNOCOMPILE_shouldCompile(File testCaseJavaFile) {
         assertCompiles(testCaseJavaFile);
     }
 
@@ -66,39 +56,5 @@ class ProcessorTestFilesCompileTest {
                         tc ->
                                 ProcessorTestHelper.isStandaloneCompilableTestFile(
                                         tc.nonCompliantFile));
-    }
-
-    private static void assertCompiles(File javaFile) throws IOException {
-        var compileResults = compile(javaFile);
-        boolean compileSuccessful = compileResults.getLeft();
-        String diagnosticsOutput = compileResults.getRight();
-
-        assertTrue(compileSuccessful, diagnosticsOutput);
-    }
-
-    /** Returns a pair (compileSuccess, diagnosticsOutput) */
-    private static Pair<Boolean, String> compile(File javaFile) {
-        // inspired comment by user GETah on StackOverflow: https://stackoverflow.com/a/8364016
-
-        var compiler = ToolProvider.getSystemJavaCompiler();
-        assertThat(
-                "System does not have a Java compiler, please run test suite with a JDK",
-                compiler,
-                notNullValue());
-
-        var diagnostics = new DiagnosticCollector<JavaFileObject>();
-        var fileManager = compiler.getStandardFileManager(diagnostics, null, null);
-        var compilationUnits =
-                fileManager.getJavaFileObjectsFromStrings(List.of(javaFile.getAbsolutePath()));
-        var task = compiler.getTask(null, fileManager, diagnostics, null, null, compilationUnits);
-
-        boolean success = task.call();
-
-        String diagnosticsOutput =
-                diagnostics.getDiagnostics().stream()
-                        .map(Diagnostic::toString)
-                        .collect(Collectors.joining(System.lineSeparator()));
-
-        return Pair.of(success, diagnosticsOutput);
     }
 }
